Add a faster rdf script
I tested this on 401 frames with 10479 atoms total and the run time was 1.5 seconds/file giving a speed enhancement of more than 300 times compared to the original. 
Note: after adding the Numba function, I noticed there are some differences in the values calculated as compared to Maxim's original code, but will check what is causing this and edit it accordingly.